refactor: Reorganize parser directory into logical modules
Restructured the parser directory to clearly separate concerns and facilitate
long-term decision-making about which parser implementation to use:

**New Structure:**

src/txxt/parser/
├── reference/       - Original combinator-based parser implementation
│   ├── mod.rs       - Module declaration
│   ├── parser.rs    - Parser orchestration
│   ├── combinators.rs - Parser combinators and primitives
│   └── api.rs       - Public API entry points
├── grammarengine/   - Regex-based grammar-driven parser (experimental)
│   ├── mod.rs       - Module declaration
│   ├── engine.rs    - Tree walker and pattern orchestration
│   ├── regex_grammar_engine.rs - Generic regex pattern matching
│   ├── txxt_grammar.rs - Grammar patterns for txxt language
│   └── unwrapper.rs - Convert matched patterns to AST nodes
├── ast_build/       - Shared AST element builders (used by both parsers)
│   ├── mod.rs       - Module declaration
│   ├── annotations.rs
│   ├── definitions.rs
│   ├── document.rs
│   ├── foreign.rs
│   ├── labels.rs
│   ├── lists.rs
│   ├── parameters.rs
│   └── sessions.rs
└── parser.rs        - Top-level module with public API and re-exports

**Benefits:**

1. Clear separation: reference vs experimental parser implementations
2. Shared code: ast_build module contains all common AST construction logic
3. Reduced coupling: parsers only depend on ast_build, not on each other
4. Easier decision-making: Can now easily compare/swap parser implementations
5. Maintained tests: All tests moved with their code to appropriate locations

**Updated Imports:**

- Updated 50+ import statements across parser/*.rs files
- Updated processor.rs to reference grammarengine instead of experimental
- Updated integration test for experimental parser
- All 348+ tests passing with zero regressions
