+++ src/txxt/parser/ast_build/foreign.rs
@@ -11,12 +11,12 @@ use crate::txxt::ast::{
     Annotation, ContentItem, ForeignBlock, Label, Location, Paragraph, TextContent,
 };
 use crate::txxt::lexer::Token;
-use crate::txxt::parser::combinators::{
+use crate::txxt::parser::reference::combinators::{
     byte_range_to_location, compute_byte_range_bounds, compute_location_from_locations,
     extract_text_from_locations, text_line, token,
 };
-use crate::txxt::parser::elements::annotations::{annotation_header, AnnotationHeader};
-use crate::txxt::parser::elements::definitions::definition_subject;
+use crate::txxt::parser::ast_build::annotations::{annotation_header, AnnotationHeader};
+use crate::txxt::parser::ast_build::definitions::definition_subject;
 
 /// Type alias for token with location
 type TokenLocation = (Token, Range<usize>);
@@ -157,7 +157,7 @@ pub(crate) fn foreign_block(
 #[cfg(test)]
 mod tests {
     use crate::txxt::lexer::lex;
-    use crate::txxt::parser::api::parse;
+    use crate::txxt::parser::reference::api::parse;
     use crate::txxt::processor::txxt_sources::TxxtSources;
 
     #[test]
