+++ src/txxt/parser/ast_build/lists.rs
@@ -10,7 +10,7 @@ use std::sync::Arc;
 
 use crate::txxt::ast::{ContentItem, List, ListItem, Location, TextContent};
 use crate::txxt::lexer::Token;
-use crate::txxt::parser::combinators::{
+use crate::txxt::parser::reference::combinators::{
     aggregate_locations, byte_range_to_location, compute_location_from_locations,
     extract_tokens_to_text_and_location, is_text_token, token,
 };
@@ -104,7 +104,7 @@ where
 #[cfg(test)]
 mod tests {
     use crate::txxt::lexer::lex;
-    use crate::txxt::parser::api::parse;
+    use crate::txxt::parser::reference::api::parse;
     use crate::txxt::processor::txxt_sources::TxxtSources;
     use crate::txxt::testing::assert_ast;
